moved UI Folder into Components renamed Components for better Understanding
