+++ src/components/ImgGrids/ImageGridMasonry.js
@@ -1,5 +1,5 @@
 import React from "react";
-import useFirestore from "../hooks/useFirestore";
+import useFirestore from "../../hooks/useFirestore";
 import { ImageList, ImageListItem } from "@mui/material";
 
 const ImageGridMasonry = ({ onSetImg, onSetBackdrop, store }) => {
